requirements.txt
langchain
langchain-core
langchain-google-genai
langgraph
python-dotenv
pydantic
streamlit
PyMuPDF
pytesseract
Pillow
